+++ example/example.g.dart
@@ -15,4 +15,21 @@ class MyExampleClassDecorator implements MyExampleClass {
   void foo() {
     myExampleClass.foo();
   }
+
+  @override
+  String toString() {
+    return myExampleClass.toString();
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is MyExampleClassDecorator) {
+      return myExampleClass == other.myExampleClass;
+    }
+    return myExampleClass == other;
+  }
+
+  @override
+  int get hashCode => myExampleClass.hashCode;
 }