+++ example/readme_example.g.dart
@@ -23,4 +23,21 @@ class MyServiceDecorator implements MyService {
 
   @override
   String get name => myService.name;
+
+  @override
+  String toString() {
+    return myService.toString();
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is MyServiceDecorator) {
+      return myService == other.myService;
+    }
+    return myService == other;
+  }
+
+  @override
+  int get hashCode => myService.hashCode;
 }