+++ test/edge_cases/object_methods_test.dart
@@ -45,6 +45,6 @@ const _expectedAnnotatedTests = {
 };
 
 const _expectedAnnotatedTestsWithReversedDefaults = {
-  'OverridesDefaultIncludedMethods',
-  'OverridesDefaultExcludedMethods',
+  'OverridesDefaultIncludedMethodsReversed',
+  'OverridesDefaultExcludedMethodsReversed',
 };