+++ test/edge_cases/object_methods_test_library.dart
@@ -1,7 +1,6 @@
 import 'package:decorator_annotation/decorator_annotation.dart';
 import 'package:source_gen_test/annotations.dart';
 
-// Empty class - should generate minimal decorator
 @ShouldGenerate(r'''
 class OverridesDefaultIncludedMethodsDecorator
     implements OverridesDefaultIncludedMethods {
@@ -18,6 +17,11 @@ class OverridesDefaultIncludedMethodsDecorator
 
   @override
   bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is OverridesDefaultIncludedMethodsDecorator) {
+      return overridesDefaultIncludedMethods ==
+          other.overridesDefaultIncludedMethods;
+    }
     return overridesDefaultIncludedMethods == other;
   }
 
@@ -26,18 +30,8 @@ class OverridesDefaultIncludedMethodsDecorator
 }
 ''')
 @Decorator()
-class OverridesDefaultIncludedMethods {
-  @override
-  String toString() => 'OverridesDefaultIncludedMethods';
+class OverridesDefaultIncludedMethods {}
 
-  @override
-  bool operator ==(Object other) => other is OverridesDefaultIncludedMethods;
-
-  @override
-  int get hashCode => 0;
-}
-
-// Empty class - should generate minimal decorator
 @ShouldGenerate(r'''
 class OverridesDefaultExcludedMethodsDecorator
     implements OverridesDefaultExcludedMethods {
@@ -46,6 +40,24 @@ class OverridesDefaultExcludedMethodsDecorator
   OverridesDefaultExcludedMethodsDecorator({
     required this.overridesDefaultExcludedMethods,
   });
+
+  @override
+  String toString() {
+    return overridesDefaultExcludedMethods.toString();
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is OverridesDefaultExcludedMethodsDecorator) {
+      return overridesDefaultExcludedMethods ==
+          other.overridesDefaultExcludedMethods;
+    }
+    return overridesDefaultExcludedMethods == other;
+  }
+
+  @override
+  int get hashCode => overridesDefaultExcludedMethods.hashCode;
 }
 ''')
 @Decorator()