+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+- Object methods are now forwarded even if they are not overridden
+  - still respects the global or local settings for overriding
+- Make decorator `==` method transparent by forwarding to the components `==` if possible
+- Fix value equality issue where 2 generated decorators with the same instance could be considered **not** equal
+
 ## 0.3.0
 - Added annotation-level forwarding support through `@Decorator()` constructor parameters
 
@@ -15,7 +21,6 @@
   - increase test coverage
 
 ## 0.1.0
-
   - Generate decorator classes using the `@Decorator()` annotation
   - Support for generics, mixins, operators, parameters, records, properties
   - forward_object_method options to control generation of `toString`, `==`, `hashCode`, `runtimeType`, and `noSuchMethod`